refactor: move base class to separate file
