+++ src/app/db/models/business_categories.py
@@ -1,4 +1,3 @@
-
 from sqlalchemy import (
     ForeignKey,
     Index,
@@ -7,7 +6,7 @@ from sqlalchemy import (
 from sqlalchemy.orm import Mapped  # type: ignore
 from sqlalchemy.orm import mapped_column, relationship
 
-from src.app.db.models import Base
+from src.app.db.models.base import Base
 from src.app.db.models.business import Business
