+++ src/app/db/models/business_working_hours.py
@@ -8,7 +8,7 @@ from sqlalchemy import (
 from sqlalchemy.orm import Mapped  # type: ignore
 from sqlalchemy.orm import mapped_column, relationship
 
-from src.app.db.models import Base
+from src.app.db.models.base import Base
 from src.app.db.models.business import Business
